Copilot: refactoring phase 7
