+++ cmd/oceano-state-manager/main.go
@@ -2,29 +2,27 @@ package main
 
 import (
 	"context"
-	"crypto/sha1"
-	"encoding/base64"
-	"encoding/binary"
-	"encoding/hex"
-	"encoding/json"
 	"flag"
-	"fmt"
-	"io"
 	"log"
-	"math"
-	"net"
 	"os"
 	"os/signal"
 	"path/filepath"
 	"regexp"
-	"strconv"
-	"strings"
 	"sync"
 	"syscall"
 	"time"
 
 	internallibrary "github.com/alemser/oceano-player/internal/library"
 )
+
+// File map for this service:
+// - main.go: config/types and process wiring
+// - shairport_metadata.go: shairport metadata pipe parsing + AirPlay state updates
+// - source_vu_monitor.go: physical source polling + VU boundary detection
+// - recognition_setup.go: recognizer composition (order + roles)
+// - recognition_coordinator.go: recognition workflow and persistence policies
+// - state_output.go: state projection, library sync, JSON writer loop
+// - track_helpers.go: track/artist normalization + cross-provider matching helpers
 
 // airplaySampleRate and airplayBitDepth are fixed transport characteristics for AirPlay/shairport-sync.
 const (
@@ -37,12 +35,6 @@ const (
 // Format: <item><type>HEX</type><code>HEX</code><length>N</length><data encoding="base64">B64</data></item>
 var itemRE = regexp.MustCompile(
 	`(?s)<item>\s*<type>([0-9a-fA-F]{8})</type>\s*<code>([0-9a-fA-F]{8})</code>\s*<length>\d+</length>\s*(?:<data encoding="base64">(.*?)</data>)?\s*</item>`,
-)
-
-var (
-	parenSuffixRE = regexp.MustCompile(`\s*[\(\[].*?[\)\]]\s*`)
-	nonWordRE     = regexp.MustCompile(`[^a-z0-9]+`)
-	wordTokenRE   = regexp.MustCompile(`[a-z0-9]+`)
 )
 
 // --- Output schema ---
@@ -251,582 +243,6 @@ func (m *mgr) markDirty() {
 	}
 }
 
-// --- Shairport-sync pipe reader ---
-
-// runShairportReader loops, opening and reading the metadata FIFO. On EOF or error,
-// it clears the AirPlay playing state and retries after a short delay.
-func (m *mgr) runShairportReader(ctx context.Context) {
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-		}
-
-		if err := m.readPipe(ctx); err != nil {
-			if ctx.Err() != nil {
-				return
-			}
-			log.Printf("shairport pipe error: %v — retrying in 2s", err)
-
-			m.mu.Lock()
-			wasPlaying := m.airplayPlaying
-			m.airplayPlaying = false
-			m.mu.Unlock()
-			if wasPlaying {
-				m.markDirty()
-			}
-
-			select {
-			case <-ctx.Done():
-				return
-			case <-time.After(2 * time.Second):
-			}
-		}
-	}
-}
-
-// readPipe opens the FIFO (may block until shairport-sync is running), reads items,
-// and returns when the pipe closes or context is cancelled.
-func (m *mgr) readPipe(ctx context.Context) error {
-	// os.Open on a FIFO blocks until a writer is present. Run in a goroutine
-	// so we can respect context cancellation.
-	type openResult struct {
-		f   *os.File
-		err error
-	}
-	ch := make(chan openResult, 1)
-	go func() {
-		f, err := os.Open(m.cfg.MetadataPipe)
-		ch <- openResult{f, err}
-	}()
-
-	var f *os.File
-	select {
-	case r := <-ch:
-		if r.err != nil {
-			return r.err
-		}
-		f = r.f
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-	defer f.Close()
-
-	log.Printf("shairport-sync metadata pipe connected")
-
-	buf := make([]byte, 0, 4096)
-	tmp := make([]byte, 16384)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-
-		n, err := f.Read(tmp)
-		if n > 0 {
-			buf = append(buf, tmp[:n]...)
-			buf = m.drainItems(buf)
-			if len(buf) > maxBufSize {
-				// Trim to last 8 KB on malformed stream to prevent unbounded growth.
-				buf = buf[len(buf)-8192:]
-			}
-		}
-		if err != nil {
-			if err == io.EOF {
-				log.Printf("shairport-sync metadata pipe closed")
-				return nil
-			}
-			return err
-		}
-	}
-}
-
-// drainItems parses all complete metadata items from buf, applies each to state,
-// and returns the unconsumed remainder.
-func (m *mgr) drainItems(buf []byte) []byte {
-	locs := itemRE.FindAllSubmatchIndex(buf, -1)
-	if len(locs) == 0 {
-		return buf
-	}
-
-	for _, loc := range locs {
-		typeHex := string(buf[loc[2]:loc[3]])
-		codeHex := string(buf[loc[4]:loc[5]])
-
-		var rawData []byte
-		if loc[6] >= 0 {
-			b64 := strings.TrimSpace(string(buf[loc[6]:loc[7]]))
-			if decoded, err := base64.StdEncoding.DecodeString(b64); err == nil {
-				rawData = decoded
-			}
-		}
-
-		m.applyItem(decodeTag(typeHex), decodeTag(codeHex), rawData)
-	}
-
-	return buf[locs[len(locs)-1][1]:]
-}
-
-// applyItem updates internal state based on one shairport-sync metadata item.
-// Mirrors the OceanoClient._apply_item() logic from oceano-now-playing.
-func (m *mgr) applyItem(itemType, code string, data []byte) {
-	strVal := strings.TrimSpace(string(data))
-
-	if itemType == "core" {
-		if strVal == "" {
-			return
-		}
-		var changed bool
-		m.mu.Lock()
-		switch code {
-		case "minm": // track title
-			changed = m.title != strVal
-			m.title = strVal
-		case "asar": // artist
-			changed = m.artist != strVal
-			m.artist = strVal
-		case "asal": // album
-			changed = m.album != strVal
-			m.album = strVal
-		}
-		m.mu.Unlock()
-		if changed {
-			if m.cfg.Verbose {
-				log.Printf("AirPlay: %s = %q", code, strVal)
-			}
-			m.markDirty()
-		}
-		return
-	}
-
-	if itemType != "ssnc" {
-		return
-	}
-
-	switch code {
-	case "pbeg": // play session begin
-		m.mu.Lock()
-		m.airplayPlaying = true
-		m.seekMS = 0
-		m.durationMS = 0
-		m.seekUpdatedAt = time.Now()
-		m.mu.Unlock()
-		m.markDirty()
-		log.Printf("AirPlay: play begin")
-
-	case "prsm": // play resume
-		m.mu.Lock()
-		wasPlaying := m.airplayPlaying
-		m.airplayPlaying = true
-		m.mu.Unlock()
-		if !wasPlaying {
-			m.markDirty()
-		}
-		log.Printf("AirPlay: play resume")
-
-	case "pend", "pfls", "stop": // play end / flush / stop
-		m.mu.Lock()
-		wasPlaying := m.airplayPlaying
-		m.airplayPlaying = false
-		m.mu.Unlock()
-		if wasPlaying {
-			m.markDirty()
-			log.Printf("AirPlay: stopped (%s)", code)
-		}
-
-	case "prgr": // progress: "start/current/end" as 32-bit RTP ticks at 44100 Hz
-		parts := strings.Split(strVal, "/")
-		if len(parts) < 3 {
-			return
-		}
-		start, e1 := strconv.ParseInt(parts[0], 10, 64)
-		current, e2 := strconv.ParseInt(parts[1], 10, 64)
-		end, e3 := strconv.ParseInt(parts[2], 10, 64)
-		if e1 != nil || e2 != nil || e3 != nil {
-			return
-		}
-		seekMS := max(ticksDiff(start, current)*1000/44100, 0)
-		durMS := max(ticksDiff(start, end)*1000/44100, 0)
-
-		m.mu.Lock()
-		m.airplayPlaying = true
-		m.seekMS = seekMS
-		m.durationMS = durMS
-		m.seekUpdatedAt = time.Now()
-		m.mu.Unlock()
-		m.markDirty()
-
-	case "PICT": // embedded album artwork (JPEG/PNG bytes)
-		if len(data) == 0 {
-			return
-		}
-		path := m.saveArtwork(data)
-		if path == "" {
-			return
-		}
-		m.mu.Lock()
-		m.artworkPath = path
-		m.mu.Unlock()
-		m.markDirty()
-		log.Printf("AirPlay: artwork saved → %s", path)
-	}
-}
-
-// ticksDiff returns the difference between two 32-bit RTP timestamps, handling wraparound.
-func ticksDiff(start, end int64) int64 {
-	if end >= start {
-		return end - start
-	}
-	return (1 << 32) - start + end
-}
-
-// saveArtwork writes raw image bytes to a content-addressed file and returns the path.
-// If the file already exists (same content hash), it is reused without rewriting.
-func (m *mgr) saveArtwork(data []byte) string {
-	h := sha1.Sum(data)
-	path := filepath.Join(m.cfg.ArtworkDir, fmt.Sprintf("oceano-artwork-%x.jpg", h[:8]))
-	if _, err := os.Stat(path); err == nil {
-		return path
-	}
-	if err := os.WriteFile(path, data, 0o644); err != nil {
-		log.Printf("failed to save artwork: %v", err)
-		return ""
-	}
-	return path
-}
-
-// --- Source detector watcher ---
-
-// runSourceWatcher polls the source detector JSON file every 2 seconds.
-// Changes are rare (user switches from vinyl to CD), so polling is sufficient.
-func (m *mgr) runSourceWatcher(ctx context.Context) {
-	ticker := time.NewTicker(2 * time.Second)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			m.pollSourceFile()
-		}
-	}
-}
-
-func (m *mgr) pollSourceFile() {
-	data, err := os.ReadFile(m.cfg.SourceFile)
-	if err != nil {
-		return
-	}
-	var det detectorOutput
-	if err := json.Unmarshal(data, &det); err != nil {
-		return
-	}
-	src := det.Source
-	if src == "" {
-		src = "None"
-	}
-
-	m.mu.Lock()
-	changed := m.physicalSource != src
-	newSession := false
-	if src == "Physical" {
-		// A new session is one where we've been silent longer than the idle delay —
-		// meaning the user stopped the record and started a new one, not just a
-		// gap between tracks or brief oscillation around the threshold.
-		if m.lastPhysicalAt.IsZero() || time.Since(m.lastPhysicalAt) > m.cfg.IdleDelay {
-			newSession = true
-		}
-		m.lastPhysicalAt = time.Now()
-	}
-	if newSession {
-		m.recognitionResult = nil
-		m.physicalArtworkPath = ""
-		m.physicalFormat = ""
-		m.shazamContinuityReady = false
-		m.lastContinuityMismatchAt = time.Time{}
-		m.lastContinuityMismatchFrom = ""
-		m.lastContinuityMismatchTo = ""
-	}
-	needsTrigger := src == "Physical" && m.recognitionResult == nil
-	m.physicalSource = src
-	m.mu.Unlock()
-
-	if changed {
-		log.Printf("physical source: %s", src)
-		m.markDirty()
-	} else if src == "Physical" {
-		// Even without a state change, mark dirty so the idle screen wakes
-		// promptly if the writer missed a previous Physical transition.
-		m.markDirty()
-	}
-
-	if needsTrigger {
-		select {
-		case m.recognizeTrigger <- recognizeTrigger{isBoundary: false}:
-		default:
-		}
-	}
-}
-
-// --- State builder and writer ---
-
-// buildState merges AirPlay and physical source state into the output schema.
-// Source priority: AirPlay (active session) > physical detector (Vinyl/CD) > None.
-//
-// Idle delay: when physical audio stops, the last track is kept visible for
-// IdleDelay seconds before switching to the idle screen. This covers the normal
-// gap between tracks on a record without blanking the display.
-func (m *mgr) buildState() PlayerState {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-
-	source := "None"
-	state := "stopped"
-
-	// physicalActive is true either when audio is currently detected, or when
-	// it stopped recently enough to still be within the idle delay window.
-	physicalActive := m.physicalSource == "Physical" ||
-		(!m.lastPhysicalAt.IsZero() && time.Since(m.lastPhysicalAt) < m.cfg.IdleDelay)
-
-	switch {
-	case m.airplayPlaying:
-		// Streaming source takes priority — physical media detection is ignored
-		// when AirPlay (or future Bluetooth/UPnP) is active.
-		source = "AirPlay"
-		state = "playing"
-	case physicalActive:
-		source = "Physical"
-		state = "playing"
-	}
-
-	var track *TrackInfo
-	displaySource := source
-	physFmt := "" // populated when Physical source format is known
-	if source == "Physical" {
-		// physicalFormat persists across track boundaries so source stays
-		// "CD"/"Vinyl" even when recognitionResult is nil between tracks.
-		fmtStr := m.physicalFormat
-		if m.recognitionResult != nil && m.recognitionResult.Format != "" {
-			fmtStr = m.recognitionResult.Format
-		}
-		switch strings.ToLower(strings.TrimSpace(fmtStr)) {
-		case "cd":
-			displaySource = "CD"
-			physFmt = "CD"
-		case "vinyl":
-			displaySource = "Vinyl"
-			physFmt = "Vinyl"
-		}
-	}
-
-	switch source {
-	case "AirPlay":
-		track = &TrackInfo{
-			Title:         m.title,
-			Artist:        m.artist,
-			Album:         m.album,
-			DurationMS:    m.durationMS,
-			SeekMS:        m.seekMS,
-			SeekUpdatedAt: m.seekUpdatedAt.UTC().Format(time.RFC3339),
-			SampleRate:    airplaySampleRate,
-			BitDepth:      airplayBitDepth,
-			ArtworkPath:   m.artworkPath,
-		}
-	case "Physical":
-		if r := m.recognitionResult; r != nil {
-			var sampleRate, bitDepth string
-			if strings.EqualFold(strings.TrimSpace(r.Format), "cd") {
-				sampleRate = airplaySampleRate
-				bitDepth = airplayBitDepth
-			}
-			track = &TrackInfo{
-				Title:       r.Title,
-				Artist:      r.Artist,
-				Album:       r.Album,
-				SampleRate:  sampleRate,
-				BitDepth:    bitDepth,
-				ArtworkPath: m.physicalArtworkPath,
-			}
-		}
-		// track remains nil until recognition identifies the track.
-	}
-
-	return PlayerState{
-		Source:    displaySource,
-		Format:    physFmt,
-		State:     state,
-		Track:     track,
-		UpdatedAt: time.Now().UTC().Format(time.RFC3339),
-	}
-}
-
-// runVUMonitor subscribes to the VU socket from oceano-source-detector and detects
-// silence→audio transitions that signal a new track starting. On each transition it
-// sends to m.recognizeTrigger so the recognizer runs at the right moment rather than
-// on a blind timer.
-//
-// VU frames are 8 bytes each (float32 L + float32 R, little-endian) at ~21.5 Hz.
-// A track boundary is: avg RMS below silenceThreshold for silenceFrames consecutive
-// frames, followed by avg RMS above silenceThreshold for activeFrames frames.
-func (m *mgr) runVUMonitor(ctx context.Context) {
-	const (
-		silenceThreshold = float32(0.01)
-		silenceFrames    = 22 // ~1 s of silence (vinyl inter-track gaps can be < 2 s)
-		activeFrames     = 11 // ~0.5 s of audio resumption
-		retryDelay       = 5 * time.Second
-	)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		default:
-		}
-
-		conn, err := net.Dial("unix", m.cfg.VUSocket)
-		if err != nil {
-			select {
-			case <-ctx.Done():
-				return
-			case <-time.After(retryDelay):
-			}
-			continue
-		}
-
-		log.Printf("VU monitor: connected to %s", m.cfg.VUSocket)
-		m.readVUFrames(ctx, conn, silenceThreshold, silenceFrames, activeFrames)
-		conn.Close()
-
-		if ctx.Err() != nil {
-			return
-		}
-		log.Printf("VU monitor: disconnected — reconnecting in %s", retryDelay)
-		select {
-		case <-ctx.Done():
-			return
-		case <-time.After(retryDelay):
-		}
-	}
-}
-
-func (m *mgr) readVUFrames(ctx context.Context, conn net.Conn, silenceThreshold float32, silenceFrames, activeFrames int) {
-	const (
-		// Energy-change detection for track boundaries without silence gaps
-		// (e.g. vinyl with residual hum, CD crossfades).
-		//
-		// Two EMAs track the signal energy: a slow baseline (~9 s time constant)
-		// and a fast current level (~0.3 s). When the fast EMA dips well below the
-		// baseline and then recovers, it indicates a track transition.
-		energySlowAlpha      = float32(0.005) // ~200-frame / ~9 s time constant
-		energyFastAlpha      = float32(0.15)  // ~7-frame  / ~0.3 s time constant
-		energyDipRatio       = float32(0.45)  // fast < slow*0.45 → dip in progress
-		energyRecoverRatio   = float32(0.75)  // fast > slow*0.75 after dip → boundary
-		energyDipMinFrames   = 7              // dip must sustain ~0.3 s before committing
-		energyWarmupFrames   = 200            // frames before detection is active (~9 s)
-		energyChangeCooldown = 12 * time.Second
-	)
-
-	buf := make([]byte, 8)
-	silenceCount := 0
-	activeCount := 0
-	inSilence := false
-
-	// Energy-change detection state.
-	var slowEMA, fastEMA float32
-	energyFrameCount := 0 // counts only non-silent frames; resets after any boundary
-	dipCount := 0
-	hadDip := false
-	lastEnergyTrigger := time.Time{}
-
-	fireBoundaryTrigger := func(reason string) {
-		// Do not clear current metadata/artwork here: boundary detection can fire
-		// on false positives, and clearing preemptively causes UI flicker/regressions.
-		// State is cleared later only when a boundary-triggered recognition confirms no match.
-		log.Printf("VU monitor: track boundary detected (%s) — triggering recognition", reason)
-		select {
-		case m.recognizeTrigger <- recognizeTrigger{isBoundary: true}:
-		default:
-		}
-	}
-
-	go func() {
-		<-ctx.Done()
-		conn.SetDeadline(time.Now())
-	}()
-
-	for {
-		if _, err := io.ReadFull(conn, buf); err != nil {
-			return
-		}
-		left := math.Float32frombits(binary.LittleEndian.Uint32(buf[0:4]))
-		right := math.Float32frombits(binary.LittleEndian.Uint32(buf[4:8]))
-		avg := (left + right) / 2
-
-		// --- Silence-based boundary detection ---
-		if avg < silenceThreshold {
-			silenceCount++
-			activeCount = 0
-			if silenceCount >= silenceFrames && !inSilence {
-				inSilence = true
-				log.Printf("VU monitor: silence detected")
-				// Freeze energy model during silence; it will restart fresh on resumption.
-				energyFrameCount = 0
-				dipCount = 0
-				hadDip = false
-			}
-		} else {
-			activeCount++
-			silenceCount = 0
-			if inSilence && activeCount >= activeFrames {
-				inSilence = false
-				fireBoundaryTrigger("silence→audio")
-				lastEnergyTrigger = time.Now()
-			}
-		}
-
-		// --- Energy-change detection (seamless transitions without silence) ---
-		// Don't run during silence or while still in the active-count phase after silence.
-		if avg < silenceThreshold || inSilence {
-			continue
-		}
-
-		energyFrameCount++
-		if energyFrameCount == 1 {
-			slowEMA = avg
-			fastEMA = avg
-		} else {
-			slowEMA = energySlowAlpha*avg + (1-energySlowAlpha)*slowEMA
-			fastEMA = energyFastAlpha*avg + (1-energyFastAlpha)*fastEMA
-		}
-
-		if energyFrameCount < energyWarmupFrames {
-			continue // EMA not yet converged; avoid false positives on startup
-		}
-
-		if fastEMA < slowEMA*energyDipRatio {
-			dipCount++
-			if dipCount >= energyDipMinFrames {
-				hadDip = true
-			}
-		} else {
-			dipCount = 0
-			if hadDip && fastEMA > slowEMA*energyRecoverRatio {
-				hadDip = false
-				if time.Since(lastEnergyTrigger) >= energyChangeCooldown {
-					lastEnergyTrigger = time.Now()
-					energyFrameCount = 0 // restart model for the new track
-					fireBoundaryTrigger("energy-change")
-				} else {
-					log.Printf("VU monitor: energy-change boundary suppressed (cooldown active)")
-				}
-			}
-		}
-	}
-}
-
 // runRecognizer waits for triggers from m.recognizeTrigger (sent on Physical source
 // activation and on track boundaries detected by runVUMonitor) and identifies the
 // playing track via the given Recognizer. It is a no-op when rec is nil.
@@ -991,280 +407,6 @@ func (m *mgr) runShazamContinuityMonitor(ctx context.Context, shazamRec Recogniz
 		default:
 		}
 	}
-}
-
-// runLibrarySync periodically refreshes the in-memory physical track metadata
-// from the library DB. This makes UI edits visible in state.json without
-// waiting for a new recognition cycle.
-func (m *mgr) runLibrarySync(ctx context.Context, lib *internallibrary.Library) {
-	if lib == nil {
-		return
-	}
-
-	ticker := time.NewTicker(3 * time.Second)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ticker.C:
-			m.syncFromLibrary(lib)
-		}
-	}
-}
-
-// syncFromLibrary updates recognitionResult from the DB when a row exists for
-// the current track (matched by ACRID or ShazamID) and user-edited fields differ
-// from in-memory values. This makes UI edits visible in state.json without waiting
-// for a new recognition cycle — including Shazam-only tracks (no ACRID).
-func (m *mgr) syncFromLibrary(lib *internallibrary.Library) {
-	m.mu.Lock()
-	r := m.recognitionResult
-	if r == nil || (r.ACRID == "" && r.ShazamID == "") || m.physicalSource != "Physical" {
-		m.mu.Unlock()
-		return
-	}
-	acrid := r.ACRID
-	shazamID := r.ShazamID
-	m.mu.Unlock()
-
-	entry, err := lib.LookupByIDs(acrid, shazamID)
-	if err != nil || entry == nil {
-		return
-	}
-
-	m.mu.Lock()
-	changed := false
-	currentACRID := ""
-	currentShazamID := ""
-	if m.recognitionResult != nil {
-		currentACRID = m.recognitionResult.ACRID
-		currentShazamID = m.recognitionResult.ShazamID
-	}
-	// Match by whichever ID is available — same logic as LookupByIDs.
-	entryMatchesCurrent := (acrid != "" && currentACRID == acrid) ||
-		(shazamID != "" && currentShazamID == shazamID)
-	if m.recognitionResult != nil && entryMatchesCurrent {
-		if m.recognitionResult.Title != entry.Title {
-			m.recognitionResult.Title = entry.Title
-			changed = true
-		}
-		if m.recognitionResult.Artist != entry.Artist {
-			m.recognitionResult.Artist = entry.Artist
-			changed = true
-		}
-		if m.recognitionResult.Album != entry.Album {
-			m.recognitionResult.Album = entry.Album
-			changed = true
-		}
-		if m.recognitionResult.Format != entry.Format {
-			m.recognitionResult.Format = entry.Format
-			if f := strings.ToLower(strings.TrimSpace(entry.Format)); f == "cd" || f == "vinyl" {
-				m.physicalFormat = entry.Format
-			}
-			changed = true
-		}
-		if m.physicalArtworkPath != entry.ArtworkPath {
-			m.physicalArtworkPath = entry.ArtworkPath
-			changed = true
-		}
-	}
-	m.mu.Unlock()
-
-	if changed {
-		if m.cfg.Verbose {
-			log.Printf("library sync: metadata updated for acrid=%s shazam_id=%s", acrid, shazamID)
-		}
-		m.markDirty()
-	}
-}
-
-// runWriter consumes change notifications and atomically writes the state JSON file.
-// It also re-evaluates state on a 5-second tick so that the idle delay expiry is
-// reflected in the output file without waiting for another event.
-// Runs in the main goroutine.
-func (m *mgr) runWriter(ctx context.Context) {
-	ticker := time.NewTicker(5 * time.Second)
-	defer ticker.Stop()
-
-	write := func() {
-		ps := m.buildState()
-		if err := writeStateFile(m.cfg.OutputFile, ps); err != nil {
-			log.Printf("failed to write state: %v", err)
-			return
-		}
-		if m.cfg.Verbose {
-			log.Printf("state written: source=%s state=%s", ps.Source, ps.State)
-		}
-	}
-
-	for {
-		select {
-		case <-ctx.Done():
-			_ = writeStateFile(m.cfg.OutputFile, PlayerState{
-				Source:    "None",
-				State:     "stopped",
-				UpdatedAt: time.Now().UTC().Format(time.RFC3339),
-			})
-			return
-		case <-m.notify:
-			write()
-		case <-ticker.C:
-			// Re-evaluate periodically so idle delay expiry is written promptly.
-			// Also write once just after the window expires so state=stopped is pushed.
-			m.mu.Lock()
-			physNone := m.physicalSource != "Physical"
-			wasPhysical := !m.lastPhysicalAt.IsZero()
-			elapsed := time.Since(m.lastPhysicalAt)
-			inIdleWindow := physNone && wasPhysical && elapsed < m.cfg.IdleDelay
-			justExpired := physNone && wasPhysical && elapsed >= m.cfg.IdleDelay && elapsed < m.cfg.IdleDelay+10*time.Second
-			m.mu.Unlock()
-			if inIdleWindow || justExpired {
-				write()
-			}
-		}
-	}
-}
-
-func writeStateFile(path string, ps PlayerState) error {
-	b, _ := json.MarshalIndent(ps, "", "  ")
-	tmp := path + ".tmp"
-	if err := os.WriteFile(tmp, b, 0o644); err != nil {
-		return err
-	}
-	return os.Rename(tmp, path)
-}
-
-// --- Helpers ---
-
-func decodeTag(hexStr string) string {
-	b, err := hex.DecodeString(hexStr)
-	if err != nil {
-		return ""
-	}
-	return string(b)
-}
-
-func normalizeTrackPart(s string) string {
-	s = strings.ToLower(strings.TrimSpace(s))
-	s = parenSuffixRE.ReplaceAllString(s, " ")
-	s = nonWordRE.ReplaceAllString(s, "")
-	return s
-}
-
-func artistTokenSet(s string) map[string]struct{} {
-	s = strings.ToLower(strings.TrimSpace(s))
-	s = parenSuffixRE.ReplaceAllString(s, " ")
-	tokens := wordTokenRE.FindAllString(s, -1)
-	ignore := map[string]struct{}{
-		"the": {}, "and": {}, "feat": {}, "featuring": {},
-		"group": {}, "band": {}, "orchestra": {}, "ensemble": {},
-		"quartet": {}, "trio": {}, "choir": {},
-	}
-	set := make(map[string]struct{}, len(tokens))
-	for _, token := range tokens {
-		if _, skip := ignore[token]; skip {
-			continue
-		}
-		set[token] = struct{}{}
-	}
-	return set
-}
-
-func tokenSetSubset(a, b map[string]struct{}) bool {
-	if len(a) == 0 || len(a) > len(b) {
-		return false
-	}
-	for token := range a {
-		if _, ok := b[token]; !ok {
-			return false
-		}
-	}
-	return true
-}
-
-func artistsEquivalent(a, b string) bool {
-	aNorm := normalizeTrackPart(a)
-	bNorm := normalizeTrackPart(b)
-	if aNorm == "" || bNorm == "" {
-		return false
-	}
-	if aNorm == bNorm {
-		return true
-	}
-	aTokens := artistTokenSet(a)
-	bTokens := artistTokenSet(b)
-	if len(aTokens) == 0 || len(bTokens) == 0 {
-		return false
-	}
-	if len(aTokens) == len(bTokens) && tokenSetSubset(aTokens, bTokens) {
-		return true
-	}
-	shorter := aTokens
-	longer := bTokens
-	if len(shorter) > len(longer) {
-		shorter, longer = longer, shorter
-	}
-	return len(shorter) >= 2 && tokenSetSubset(shorter, longer)
-}
-
-func tracksEquivalent(aTitle, aArtist, bTitle, bArtist string) bool {
-	aT := normalizeTrackPart(aTitle)
-	bT := normalizeTrackPart(bTitle)
-	if aT == "" || bT == "" {
-		return false
-	}
-	return aT == bT && artistsEquivalent(aArtist, bArtist)
-}
-
-func sameTrackByProviderIDs(a, b *RecognitionResult) bool {
-	if a == nil || b == nil {
-		return false
-	}
-	if a.ACRID != "" && b.ACRID != "" {
-		return a.ACRID == b.ACRID
-	}
-	if a.ShazamID != "" && b.ShazamID != "" {
-		return a.ShazamID == b.ShazamID
-	}
-	return tracksEquivalent(a.Title, a.Artist, b.Title, b.Artist)
-}
-
-func canonicalTrackKey(r *RecognitionResult) string {
-	if r == nil {
-		return ""
-	}
-	if r.ACRID != "" {
-		return "acrid:" + r.ACRID
-	}
-	if r.ShazamID != "" {
-		return "shazam:" + r.ShazamID
-	}
-	return "meta:" + normalizeTrackPart(r.Title) + "|" + normalizeTrackPart(r.Artist)
-}
-
-func chooseConfirmationResult(
-	primaryName string,
-	primaryRes *RecognitionResult,
-	primaryErr error,
-	confirmName string,
-	confirmRes *RecognitionResult,
-	confirmErr error,
-) (*RecognitionResult, error, string) {
-	if confirmErr == nil && confirmRes != nil {
-		return confirmRes, nil, confirmName
-	}
-	if primaryErr == nil && primaryRes != nil {
-		return primaryRes, nil, primaryName
-	}
-	if confirmErr != nil {
-		return nil, confirmErr, confirmName
-	}
-	if primaryErr != nil {
-		return nil, primaryErr, primaryName
-	}
-	return nil, nil, ""
 }
 
 // --- Main ---
